refactor: centralize reference matrix serialization
